Move out format funcs from cli core
